refactor: Organize modules into layers to simplify dependencies (#3)
Introducing a layered architecture, in an effort to better represent
dependencies between the 18 top-level modules.
`common` -> `lang` -> `semantics` / `interface` -> `pass` / `interp` -> `runner`
is the simplified dependency graph.
* refactor(runtime-dynamic): Move type substitution back into `Runtime_dynamic.Typ`
* refactor: Group modules into layers
* refactor: Merge `domain` and `util` into `common`
* refactor(builtins): Move `builtins` out of `eval_il`

Everywhere one imports language modules such as `El` and `Sl`, one
needs access to the type definitions. Thus, following the structure
introduced in `Il`, other language modules adopt this design and
include their type definitions automatically.
Also, `Il` no longer includes `Effects`, this should instead be imported explicitly. `ast.ml` is renamed `types.ml` to better represent their role.
* refactor(il): Remove redundant `core` and move to `il`
* refactor(lang): Make each `Lang` module carry type definitions

Handlers are registered on the client-side for better composability.
This comes with the cost that the API cannot ensure effect-throwing
functions are enclosed in a handler, which must be addressed in the
future. Created `eval_il` and `eval_sl` which are the atomic pipeline functions,
and `eval_il_p4_typechecker` which composes the functions.
Corresponding changes are made in `bin` and `test` scripts.
* refactor(runner): Call handlers from client-side
* refactor(runner): Separate `interp` functions into atomic and composed

Test files are also reorganized, and documentation is updated for struct/interp commands.
* test: Update expects
* test: Re-organize test files
* docs: Document binary commands